requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
discord.py==1.5.0
entrypoints==0.3
gevent==20.5.0
greenlet==0.4.15
grequests==0.6.0
gTTS==2.1.1
gTTS-token==1.1.3
hypixel==0.7.9
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.9
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==7.2.0
multidict==4.7.5
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.18.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.7.2
PyAudio @ file:///C:/Users/Eleve/Downloads/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser==2.20
pygame==1.9.6
Pygments==2.4.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.6
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.12.0
pytz==2019.3
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.23.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.9
vboxapi==1.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
websockets==8.1
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==0.6.0
